+++ src/backend/compare/compare.ts
@@ -9,7 +9,7 @@ import type {
 } from '../../views/view-types.js';
 import { respondProjectNotFound } from '../common/standard-responses.js';
 import { dbConfig, refreshSecret } from '../util.js';
-import { prepareTemplate, processTemplate } from '../../templates.js';
+import { prepareTemplate, processTemplate } from '../templates.js';
 import { deleteReport, renderCompare, renderCompareNew } from './report.js';
 import * as dataFormatters from '../../shared/data-format.js';
 import * as viewHelpers from '../../views/helpers.js';
